refactored the abstract entity test and moved the test entity generation into its own class
also implemented a saver interface as I realised it was missing
