--- src/Entity/AbstractTest.php
@@ -1,10 +1,0 @@
-<?php declare(strict_types=1);
-
-namespace EdmondsCommerce\DoctrineStaticMeta\Entity;
-
-use PHPUnit\Framework\TestCase;
-
-abstract class AbstractTest extends TestCase
-{
-
-}
